+++ wgpu-hal/src/vulkan/conv.rs
@@ -247,6 +247,15 @@ pub fn map_texture_usage(usage: wgt::TextureUses) -> vk::ImageUsageFlags {
     if usage.contains(wgt::TextureUses::COLOR_TARGET) {
         flags |= vk::ImageUsageFlags::COLOR_ATTACHMENT;
     }
+    if usage.contains(wgt::TextureUses::RESOURCE)
+        && usage.intersects(
+            wgt::TextureUses::COLOR_TARGET
+                | wgt::TextureUses::DEPTH_STENCIL_READ
+                | wgt::TextureUses::DEPTH_STENCIL_WRITE,
+        )
+    {
+        flags |= vk::ImageUsageFlags::INPUT_ATTACHMENT;
+    }
     if usage
         .intersects(wgt::TextureUses::DEPTH_STENCIL_READ | wgt::TextureUses::DEPTH_STENCIL_WRITE)
     {